% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgdal_polygonize.R
\name{rgdal_polygonize}
\alias{rgdal_polygonize}
\title{GDALPolygonize in R}
\usage{
rgdal_polygonize(raster, folder, layer, field)
}
\arguments{
\item{raster}{name/location of raster file to polygonize. Any driver included in GDAL should work.}

\item{folder}{name/location of the shape file to write. File extension should not be included.}

\item{layer}{name of the layer (.shp file) to write.}

\item{field}{name of the field that will keep the raster values.}
}
\value{
integer 0
}
\description{
The function calls the C GDALPolygonize routine (using C++). It does not overwrite. It uses always 8 point connectedness. It returns always a file in "ESRI Shapefile" format.
}
\examples{
NULL
}
