% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGDALPolygonize.R
\name{RGDALPolygonize}
\alias{RGDALPolygonize}
\title{GDALPolygonize}
\usage{
RGDALPolygonize(input, output)
}
\arguments{
\item{input}{input name of raster file to polygonize}

\item{output}{output name of the shape file to write}
}
\value{
integer 0
}
\description{
GDALPolygonize is not available in rgdal. Moreover, it is only available in C or Python. The function implements GDALPolygonize in the R environment
}
\examples{
NULL
}
